refactor: Move out relations code into own source file
